api/requirements.txt
Flask>=3.0.0
psycopg2-binary>=2.9.9
gunicorn
